JMH gives... counter-intuitive benchmark results
JMH needs to be in its own module, but it doesn't need to be distributed
with any end-user-usable libraries. For small maps, JMH gives the very
odd result of DijkstraMap.findPath() being slightly faster than
DijkstraMap.scan(), even though the "slower" one should always be called
by the "faster" one. This seems to happen on 60x60 and 30x30 maps, but
the trend is reversed on large maps (120x120).