Quick test to verify that Hilbert really is ideal
The alternative to using a lookup table for the 256x256 Hilbert Curve is
using Z-Codes, which are the same as Morton numbers and can be computed
quickly without a lookup table. But, the tests show that even though the
optimal cases are similar (Hilbert still does better), the worst-case is
much worse for Z-Codes in terms of memory usage, around 40% more.