squidlib-gdx should be fully Android-compatible!
Last use of AWT (in SquidInput to get Caps Lock state) is gone; since we
can't track Caps Lock in any sort of cross-platform way, we won't track
it at all. The static state in DefaultResources should be handled in a
way that Android is OK with, and @smelc's request for a "super good
looking complete font" has not gone unnoticed. We now have a Latin,
Greek, and Cyrillic-supporting bitmap font in GDX that still has my
beloved box drawing characters, plus a bunch of other nice things. It
also is very pretty!