CoordPacker has new useful methods!
You can count the number of off or on cells in a packed array without
unpacking, you can get the whole map as "on" and anything not on the map
as "off" (useful for intersection), and you can do random choosing of
one element or random sampling of possibly many elements! Also pack()
works on char[][] now if you give it a char to encode as "on"; this lets
us build packed arrays for walls, floors, water, etc.