Small, important fix to CoordPacker in JARs
Previously, CoordPacker relied on pre-computed lookup tables to avoid
calculating the full curve's x, y, and reverse lookups for all 65,536
items. That's still not bad to compute at startup though, and this
avoids reading from disk -- which turns out to be vital when JARs have
encodings that mess up binary files. It is possible that 3.0.0-b1 will
not have usable CoordPacker code because it uses LUTs, those can't be
loaded properly from some JARs, and I don't know if the ones I published
have that problem.