GDX UI classes; use ColorCenters more (BREAKING)
This should really make using filters easier. This breaks the API for
SquidLayers, since it removes the lightness panel (which was
translucent) entirely and instead uses the int[][] of lightnesses (that
was already present) to directly change and cache colors. I have no idea
on performance implications, but one less panel should be good, and
since the SquidColorCenters are used heavily now, filters are easier to
add in.