Back to normal, HDR is gone! KAPUT!
Other changes are still intact, but should be, with some exceptions,
unambiguous improvements. Filters are still in, and work on regular GDX
Colors now. A Filter-capable SquidColorCenter is used to get colors for
lightened areas of a SquidLayers, and you can set the color used to
lighten now instead of always using SColor.CREAM (the color also can be
set in a SquidPanel). The new dungeon generators haven't been reverted
or anything, so, hooray!