Fix build-stopping issues, continue on cover work
FOVCacheTest crashed with an error due to the behavior changing in
PoissonDisk (which it uses to place lights, I think), and that's now
fixed. The annoying JDK 8 Javadoc linting that doesn't play nicely with
IntelliJ's automatic formatting (which inserts the illegal <p />, oh how
horrible, we must stop the build...) has been disabled.