Temporary commit to keep some work in history
So the best general-purpose compressed bitset lib I could find still
uses more than twice as much memory as the current specialized-purpose
compressed bitsets used by CoordPacker. GOOD TO KNOW. Specifically, the
average for an "Appropriate" map with RoaringBitmap used 265% of the
memory of a short[] with CoordPacker, and the average for an
"Approaching Worst-case" map used 268% of the memory.