BasicDemo, update GDX, test scope for GDX platform
BasicDemo should be a nice easy way to see a good amount of SquidLib.
We're updated to libGDX 1.7.1 and a closer-to-modern version of Maven
while we're at it.
Changing GDX dependency to test scope means users on Android won't need
to download the useless (for them) LWJGL backend as a dependency to play
a game, but developers will need to specify which platform they want for
a project. This last part should be made easier by the customization I'm
doing on LibGDX's setup tool to handle automatic setup for lib users.