Fixed Poisson problems, changed grass/water gen
Due to the way MultiSpill works, it was making even fairly low
percentages fill large actual amounts of map with water or grass. This
is "remedied" now by making grass and water fill 2/3 as aggressively as
floor, and you can still force all grass or all water (or all mixed I
think) by using 150 as a percentage (this is documented).