Change KissSpill to Splash, add static method
Also changes rmImpassableChar to removeImpassableChar to match naming
conventions in the rest of SquidLib and Java standard library, and adds
a constructor that initializes the impassable set without needing to
call more methods (it's handy for the static method). The static method
creates a Splash object and caches it, storing its hashCode to quickly
check if a different Set of chars is given, which means it needs to
update the cache.