Further optimization of retract and surface
When run on typical dungeon floors, using expand instead of fringe to
get the surface or retracting area means there is a much larger
contiguous space to get the intersection or difference with the original
region, which, since those operations depend on the number of contiguous
regions packed in a short[] for their performance, a roughly 2x speedup
over the last commit for expand and surface on regions like floors that
already have large contiguous spans. for things like walls, which may
have less contiguous spans, this means potentially a slight slowdown. In
either case, the code for all these is a one-liner now, and inlining the
right one may be the best option.