MonsterGen works, room-finding in CoordPacker
MonsterGen is now documented and should be barely usable for generating
descriptions, though it needs more base creatures to mix easily. There's
two added fake languages in FakeLanguageGen. New methods in CoordPacker
to get a fraction of the Coords in some packed data with an good spacing
between Coords. A split method in CoordPacker that takes a map with
separate areas and returns an array of packed data elements, one per
separate area, with documentation on how to quickly get rooms and
corridors from an arbitrary bare dungeon. It's good.