Serpent family of mapgen fixed
On specific RNG seeds, columns or rows could be given negative
positions, which made the Hilbert Curve calculations fail since they
look up positions in an array (negative indices bad!). This appears to
be fixed; there was a loop that went in reverse order and so could
subtract a high value from a position at the start of the loop, when the
position numbers were low, so this could produce negative positions. I
tested the code from issue 115 that previously produced an exception on
10,000 sequential RNG seeds, including the one that caused the problem
initially, and none of them caused an array out of bounds exception with
this fix.