Add internal->displayed char swapping.
This is in TextCellFactory, and should be useful if/when we add more
chars used for customizable purposes. An example might be swapping '£',
used for lava, for '§' if the game has toxic sludge, but no lava. This
commit also removes the badly outdated gdx_test section of the code,
which I don't know if it was being included in release jars, but it
shouldn't be there at all now that we have SquidSetup.