Sub-cell (smaller text, really) panels
These are used in EverythingDemo to show player health as an overlay on
top of an '@' and to show monsters who are alerted with '!' over them.
This came out of a good suggestion by @SquidPony to have a sort of "Loot
Sense" where valuable things sparkle, and this could definitely be used
for that effect. We also use an "internal zoom" in EverythingDemo,
drawing at double size and scaling down when rendered, to increase the
sharpness of the tiny overlays (large, blurry pixels are visible
otherwise), but this only works with distance field fonts. Also
important to note: SquidPanel now handles the distance field font
configuration itself, so SquidLayers and anything else that uses
SquidPanel should work with distance field fonts without needing code
added (you may want to remove a call to configureShader() if you have it
where it now isn't necessary).