Starting work on Placement
Currently it correctly identifies regions in a dungeon that could be
used for large flat objects, like racks of weapons or notice boards. The
goal is to have it handle more kinds of regions where things could be
placed, though speed may be an issue if many dungeons need to be made
quickly.