Troublesome code to fix height of effects
This commit was mostly trial and error for a very long time, but was
tested with a non-distance-field font at size 6x12, and a distance field
font at 3 different sizes before zoom with a message box below (that box
turns out to be relevant!). It seems to work with only minor differences
in text-to-background distance in all of them, and though black bars
appear for maybe one frame in certain squares of an effect, they're much
harder to notice, if they appear at all. I think they may be related to
SquidColorCenter, since they happen less frequently after a particular
color has been drawn before. @smelc , I'm hoping this is the commit that
fixes it!