I really hope this helps with the height issue.
I checked the help screen in EverythingDemo more carefully this time and
resolved an unrelated bug with border drawing when the unit for drawing
isn't equivalent to 1 pixel (which will happen if you use a larger font
and scale down to enhance resize-ability). If this doesn't fix the issue
I'll probably create and expose some kind of configuration to hack in an
adjustment to positions when things are different from how this expects.