Speed improvements to ChaosRNG
It should now be more reasonable for its intended use, as a primary
RandomnessSource for RNGs in builds of a game meant for competitions.
Also, it is amazing how slow SecureRandom is on its own... I am glad we
have a somewhat-secure alternative to stuff like LightRNG (which is not
meant for competitions) now.