Placement can find centers of rooms.
This required some fixes to CoordPacker!  The fringes method ignored the
eightWay parameter when it was given, and this is now fixed. Four-way
expansions and retractions also probably had a bug for larger expansion
radii (anything more than 1), and this should be fixed. The Placement
class does take a little longer to find room centers than the other
things it can find, especially on large maps, but some optimizations
allow it to avoid lengthy calculations if the center has been found
early on.