Better everything in RandomBias
RandomBias has multiple distributions that can be chosen, defaulting to
an average of the (very accurate, but weird) exponential distribution
and the (very reasonable-seeming, but limited) triangular distribution.
RandomBias allows using the class without ever needing to specify kinds
for biases now, just using the expected averages "raw" is an option.
Added some sane toString methods for RNG objects that print the state
when it's public and print a hash otherwise. Added 64-bit hashing as an
option for CrossHash. Wrote a test that graphs different RandomBias
configurations.