New Actors that change color in a loop
This doesn't interfere with existing checks for animation, since the
colors of ColorChangeImage and ColorChangeLabel are dependent on how
much time has passed (and been passed as delta to act(), where it is
stored to determine which color to use from a list). New methods in
SquidColorCenter make it easier to produce rainbow and back-and-forth
gradients, and this is demonstrated in EverythingDemo. As an aside, 650
days of streak, woo!