Major GWT compatibility improvement. Hashes change!
This commit adds a dependency to a project I've had in the works for a
few weeks: RegExodus, a regular expression library that works on all
platforms SquidLib targets, including GWT. This allows FakeLanguageGen
to be GWT compatible and any places that use it to also now become GWT
compatible. In checking for compatibility, I found that several classes
in SquidLib turned out to use methods that are not implemented in GWT,
including floatToRawIntBits and doubleToRawLongBits in CrossHash. These
changed to the compatible floatToIntBits and doubleToLongBits, but this
may change what some double or float arrays hash to on platforms other
than GWT (where they would have not compiled). SquidSetup needs to be
updated to load the dependency to RegExodus on HTML.