FakeLanguageGen improvements/fixes
A bug in FakeLanguageGen.copy() caused the vast majority of words
generated by the copied generator to be 1 syllable in length. This only
affected FakeLanguageGen instances created by addModifiers, so it showed
up somewhat rarely in testing. This is fixed now, and the syllable
counts are correct. Modifier is now a nested class in FakeLanguageGen;
it contains both a Replacer and a double for the probability that the
modifier is used. This allows things like intermittent stuttering to
work as a Modifier, and that is one of a few predefined Modifiers in the
nested class.