On-screen buttons, options for phone input
This is in the VisualInput class, which is... a doozy. It currently
needs some changes to be made when updating from SquidInput, but if you
keep using SquidInput nothing will change. VisualInput still needs heavy
JavaDocs, and a lot of things about its use aren't intuitive at first.
It does work in EverythingDemo now, where the buttons are forced even on
desktop, but normally they will be invisible and no changes will be seen
on devices with hardware keyboards (like PCs).