Key update needed for VisualUpdate to function...
...when the buttons are not displayed. Not enough places checked for
there to actually be mouse listeners for the left and right button areas
before they called a method on them. Since those areas can be null and
their listeners the same if there is a hardware keyboard available (and
buttons aren't forced, as they has been in the demo for testing), this
caused crashes when using the new VisualInput class without buttons.
This seems to be fixed, though it needs a different (smaller, simpler)
API to display the buttons.