Imitation Somali should be good for mixing
It has a particularly unfamiliar sound on its own, and when mixed with a
small percentage of other languages like Japanese and Swahili, and
possibly with modifiers like stuttering added, it can end up seeming
completely alien. It may be suitable for societies in science fiction or
fantasy that should be the most different from ones that are familiar to
the player.