Add EditRNG; simplify some other RNGs
EditRNG is meant to be more configurable than other subclasses of RNG,
without needing to resort to RandomBias as a post-processing step.
DharmaRNG and DeckRNG got some minor cleanup that may speed them up but
will change the results of seeds with those subclasses. All cases of
nextGaussian now use Math instead of the GwtIncompatible StrictMath
class; whether this makes a difference is unclear to me since the
documentation seems identical, but now nextGaussian will work on GWT.