Major improvements supporting ModularMapGen
Major fix for infinite loop bug in SectionDungeonGenerator,
functionality fix in Radius when getting the region encoded by a circle
(other radius types weren't affected). New methods in CoordPacker,
RegionMap, and GwtCompatibility. ModularMapGenerator appears to work
now! It's been tested on a small-ish and a large-ish map and it seems to
make better maps if the maps are large. There's also a preview linked in
the ModularMapGenerator javadocs.