New XoRoRNG based on xoroshiro128+ algo
David Blackman and Sebastiano Vigna recently debuted a new RNG algorithm
that is currently the fastest in Vigna's Xorshift shootout (
http://xoroshiro.di.unimi.it/ ), so naturally SquidLib needs to
implement it. It is **tied for the fastest** RandomnessSource with
LightRNG, but also has a larger period before it starts repeating
values. It does not implement StatefulRandomness, unfortunately, but
otherwise could be the preferred RNG in SquidLib for many purposes.
Being able to efficiently get the state of LightRNG is more useful from
the library perspective (finding which states cause defective procedural
generation, for instance), so the default will remain with LightRNG but
you can create your own RNG with a XoRoRNG easily.