After much work, TextPanel can resize correctly
Thanks to @smelc for a patch that got this going on its current path.
The fix turned out being very involved and completely counter-intuitive
-- don't lay the text out again at larger sizes, since the font will
think it is the same size and is rendering in a larger area. Using the
viewport to stretch the TextPanel like the rest of the text in the
"background" turns out to work well. Future commits may be able to
remove the dependency on ShapeRenderer in TextPanel; it isn't used much
now (if at all), and since it uses absolute screen coordinates to place
shapes, it really doesn't resize well. The background drawing with
ShapeRenderer has been replaced with a simple stretched pixel of the
desired color, with the pixel sprite being the same used for backgrounds
in SquidLayers (a single solid block from TextCellFactory). Also a new
version of the Noto-Sans distance field font, minimized with TinyPNG and
copied to `assets/`.