Add LanguageCipher, improve LongPeriodRNG.
LanguageCipher builds up a mapping of words to FakeLanguageGen-generated
ciphered words in some linguistic style, and can translate a whole
passage while remembering it translated, for instance, "the" to "kai",
and continuing to use that translation. Capitalization is preserved
based on the input text, and should approximately match the input text.
LongPeriodRNG now actually generates 1024 random bits when you want a
random seed, and this is useful for SquidID.