Modifiers in FakeLanguageGen can have more parts
This allows more complex modifiers to be made and passed as a single
value, like the new replacementTable() static method used to
(optionally, but by default) change the romanized imitation Hindi to
only use ASCII and Latin Extended-A characters, allowing the default
fonts to render it all around. This also changes the RegExodus
dependency to 0.1.2, which brings in some nice bugfixes and a new
feature used in the replacementTable() method. This only affects
behavior if you use FakeLanguageGen, and should (I think) compile fine
with the 0.1.1 dependency.