New iterators for regions of a SpatialMap
You can now iterate through a sub-region of a SpatialMap with minimal
allocation, especially compared to earlier methods that created a
sub-section as a new SpatialMap. Also, you can now query (but not
directly set) the current height and width of the 2D Coord cache, which
can be useful if you aren't using CoordPacker and expect expandPool()
may be called in Coord.