Decipher text with LanguageCipher, RegExodus 0.1.3
This bumps the dependency on RegExodus to 0.1.3 to get a bugfix from
there. Deciphering text allows games to give players a partial
vocabulary that they can expand over time, and translate foreign words
to familiar ones. Documentation was also added to LanguageCipher.

There are also some parts of this commit that relate to AStarSearch and
the new PlannedAStar, but neither should currently be recommended
because benchmarks show both are well more than an order of magnitude
slower than DijkstraMap (34x slower for AStarSearch and even worse for
PlannedAStar, though not by a predictable amount). AStarSearch should
get faster in a future release once the code is revisited.