Fix and speed boost in LanguageCipher
In the last commit that introduced deciphering, it was possible for a
ciphered word and a deciphered word to be the same, leading to uncertain
translations with the multiple-pass replacement in HashMap iteration
order that commit used. Now decipher does one pass with a more complex
pattern and a replacement function that looks up a word in the
vocabulary only once, when it is being replaced. This should be faster
than iteratively replacing many patterns, and behave better when input
and output overlap. This also was checked so it behaves with capitalized
vocabulary keys and values.