Dungeon generation improvements
Now if you pass a char[][] to generate() in DungeonGenerator or
SectionDungeonGenerator, it will resize the generator's width and height
to match the array instead of possibly using a larger size than is
viable (which could throw out-of-bounds exceptions).
SectionDungeonGenerator now won't generate doors in the middle of
corridors if the corridor was specified by an environment array; before,
the slight changes in width in diagonal lines used by
ModularMapGenerator made each connecting line detect as multiple rooms
and corridors, many of which would have doors between them. Better docs
in LanguageCipher.