LanguageCipher doesn't duplicate words
This changes some behavior. Word creation behavior before was completely
deterministic, where a source word always produces a target word. Now,
it is deterministic if the order words are entered is the same; if two
source words would produce the same target word, the first to be looked
up will be the same, and the second will change. This generally prevents
the problem of two common words like "hit" and "map" possibly producing
the same target word, and then decipher() being unable to discern which
was meant. The problem is still possible on extremely bad
FakeLanguageGen objects, such as languages with one vowel and one
consonant. This also updates to RegExodus 0.1.4, which fixes a problem
in word recognition that matters here.