Added MimicFill, port of ConvChain
MimicFill has a test that shows it does work as intended, it seems.
There are some small optimizations in MimicFill to avoid unnecessary
allocation of Pattern objects and avoid the general pattern of using
Func and Action from C# since Java (7) would need a more verbose way to
do the same thing. Predefined samples of 2D boolean arrays should be in
an upcoming commit. Also, some small updates to CoordPacker and Region
to allow randomized, but separated, fetching of a fraction of Coords in
packed data.