Added predefined samples in MimicFill
These are almost-exact copies of the black-and-white BMP files
distributed with ConvChain, but are transposed to make the conversion
easier. This shouldn't affect the kind of data generated, since the
transposed version should be the same as one of the flipped rotations
this used already. There's also improvements to MimicFillTest to show
what this can do.