Added GapShuffler; tweaks to RNG, StatefulRNG
GapShuffler allows getting a random ordering of elements from a
collection as an infinite Iterable, but won't ever repeat an element
(based on its index) close in order to a previous occurrence of that
element (there's a gap between occurrences of any one element). RNG now
allows any Collection in shuffle() but still returns an ArrayList, which
shouldn't break any code unless it subclassed RNG, and allows Sets and
Queues to be shuffled as well. StatefulRNG, when setRandomness() is
called with a RandomnessSource that isn't a LightRNG (which wasn't
recommended in the first place), now uses 64 random bits from that
source instead of 32, matching the existing behavior of the constructor.
This shouldn't affect much, but if your code wants consistent behavior
from a seed with StatefulRNG, setting the seed as a long is the best and
easiest way with StatefulRNG.