Start making large list of Thesaurus categories
This is the type of thing a library can be really good for, since this
would be tedious to do on your own. There are already keywords for some
medieval-political words like "dynasties" and "barony", and many
pre-made, hand-picked imitation-language words drawn from
"FakeLanguageGen's Greatest Hits", pretty much. There's a small example
in ThesaurusTest, generating phrases like, "The small state of
Dabrisazetsky in the Hegemony of Devive, ruled by Shogun Jélebi."