Can get a zigzagGradient from SCC
This is a non-linear gradient that originally used the zig-zagging path
f the Z-order (Morton) curve, but now makes a creative use of the 3D
Hilbert curve to cause a lot of changes as the gradient progresses. 3D
coordinates refer to r,g,b instead of x,y,z, and the winding path of the
Hilbert curve gives some good variation to colors along the gradient
without suddenly jumping between unrelated colors as the Z-order curve
does. Since only SquidColorCenter (currently) has gradients, only that
class has the implementation.