Significant fix for "freeze-like" bug re: animation
EverythingDemo would occasionally "lock up" and stop responding to input
while an animation was occuring. This turned out to be caused by the
hasActiveAnimations state staying above 0 even if all animations had
ended, which in turn was caused by two simultaneous animations starting
on the same cell. This seems to be fixed, along with a few other
possible causes for the bug or similar bugs. This may need to be
revisited if the bug shows up again.