Optional icons with distance field effect applied
These are in the assets/ folder along with their license, CC-BY
(commercial use OK) with several different authors to attribute.
(Because all of their work is in the compiled atlas, you should just
credit all of them, which is easier anyway. :smile: Just note that it
has been modified in addition to the authors you credit; this is already
mentioned in icons-license.txt).