Release 3.0.0-b6
It seems like enough moderately-important fixes and additions are
present here to warrant another release, and while I'm working on a
"showcase" demo for gdx-setup, I might as well fix as many relevant bugs
as possible. This commit adds a useful method in Maker to make
LinkedHashSets, and also makes an icon atlas available via
DefaultResources if the relevant files are present. This release adds a
lot more in total! See release notes.