Make "dialects" with LanguageCipher
LanguageCipher can now add a given amount to the hashcodes it gets for
source words, and then that result is used as the deterministic seed for
the RNG used to generate each word. This allows a huge variety of
dialects to be produced from the same FakeLanguageGen by using different
shift values, which can be handy if you want to have mutually
unintelligible languages that nonetheless seem similar in style or
sound.