DetailedMimic, a port of SynTex
But it isn't just for synthesizing textures! There's an example that
uses DetailedMimic to produce mimicked versions of a full-blown dungeon,
with walls, floors, water, grass, lava, and some doors, in varying sizes
given only one dungeon as input. This treats char cells in a dungeon map
like int values in an indexed-color image (indicated by the discrete
argument in DetailedMimic). Documentation should be coming soon.