Fixes to some possible bugs in dungeon gen
DenseRoomMapGenerator has some changes in the type of maps it generates;
because large areas were inaccessible on most maps before this update,
DungeonGenerator often made the panoply of dense rooms into just a
handful of connected rooms. This is partly remedied by switching
techniques for "pushing holes through" the walls, and also by making
holes (for doors or corridors) easier. DungeonGenerator and
SectionDungeonGenerator have a better check for when stairs can't be
found. DetailedMimic has a fix for some difference in porting behavior
that showed up rarely; it uses a Set instead of a List now.