New pointPath method in Serpent, various uses
There is additional code to handle these uses! DungeonUtility has a
useful ensurePath method that modifies a dungeon to carve out a valid
path with a specific char (tab is recommended) where there was one of
several blocking chars, and also returns the path it carved. Tab chars
are recommended to carve out because DungeonGenerator and
SectionDungeonGenerator will (now) both give preference to tabs first
when finding valid staircase locations (walling off anything that can't
be reached from a staircase), then will replace every tab with '.'. This
is used in DetailedMimicDungeonTest at lines 76-78. It's useful there
because the maps could have many sections that were closed-off from the
rest of the dungeon, and so were very bad maps if the staircase was
placed in a tiny enclosed room. Ensuring a path makes the maps much
better!