Decent speedup on DijkstraMap using FastUtil
In squidlib-util's squidmath package, there is now one class from
FastUtil (a data structures library supporting primitive types
efficiently), IntDoubleOrderedMap, which draws together multiple
abstract classes that FastUtil's Int2DoubleLinkedOpenHashMap class
extended into one (massive) class with many nested classes. It's very
messy code right now, having been hacked into place to work, but it
seems to yield a decent speedup on DijkstraMap calls, though it adds a
sizable amount to the dependency size since the source file is huge. For
future reference, the unmodified DijkstraMap code that uses
LinkedHashMap<Coord, Double> is in squidlib-performance (which won't be
on Maven Central), and that's how I quantified the speedup. Coord also
gained methods to decode and encode to int values losslessly for all
Coords that are within a certain range, and these are used to allow the
keys to be primitives in the IntDoubleOrderedMap implementation.