OrderedMap can replace LinkedHashMap, but...
But it's going to continue to get added features to support random
access and alternate iteration order. These are almost ready, but not
quite yet. OrderedMap was ObjectObjectOrderedMap in the last commit, and
OrderedMap from before that was removed because it didn't iterate in a
stable order for keys or entries (and stable ordering is the whole point
of using ordered Maps!).