Indexing added to OrderedMap, testing usage
You can now get an item (meaning a key, value, or key-value pair as a
Map.Entry) from an OrderedMap given its index in the iteration order,
without actually doing any iteration and in constant time. You can also,
as a result, get random items. Reordering is currently limited to
shuffling the iteration order, but work is under way to allow more
complex reordering, possibly in a simple a way as giving a list of keys
to iterate through. This only applies to OrderedMap right now, not
IntDoubleOrderedMap yet. Several other classes have been refactored
slightly to test-drive OrderedMap internally, and it seems to behave
identically in LanguageTest (where IntDoubleMap is used) and doesn't
seem to have problems in other tests (the two ordered Maps should be
identical to LHMs but I can't tell the difference relative to
LinkedHashMap on some of them).