Ordered Maps/Sets allow random access, use it!
This reduces now-unnecessary allocations of temporary arrays in several
places where an array was produced simply to get a random element from
it using a random numerical index. SpatialMap also allows random access
now via `randomIdentity()`, `randomElement()`, etc. There's also a key
fix to OrderedSet's toArray() method, which previously caused problems
when making other collections using an OrderedSet as a basis, as with
`new ArrayList(myOrderedSet)`. This was noticed in PoissonDisk and
fixed.