Important bugfix for OrderedMap, fixes demos
This was fixed already in OrderedSet and I thought it didn't show up in
OrderedMap, but there is a KeySet in OrderedMap that has the same issue.
Really it affected anything that iterated through an OrderedMap, not
just the keys, but it seems fixed now. All the demos work as before.