Fixes for Ripple (very loose) FOV, FOVCache
The Ripple FOV bug can be traced back to when it was first noted... on
September 20, 2015, [in this
commit](https://github.com/SquidPony/SquidLib/commit/ca4c31185e0ae26e2e5b8febb633749bddb0e92b),
and it was an outrageously simple fix for how long it took to get it
done... FOVCache has some issues with threads lasting too long when it's
created on the main thread (which is never a problem for JUnit, hence
why this didn't get noticed until now). The solution is to call
destroy() on your FOVCache when your game is shutting down. Finalizers
don't work here, it needs to be done by hand, it seems.