Optional threshold for diagonal-move obstacles
This applies to DijkstraMap, which has also gotten some support to
reduce the amount of garbage it produces and so (when
`blockingRequirement` is set to 0, so the slight slowdown from
additional obstacle checks is removed) the new code is a good deal
faster at finding paths, and on-par at just scanning. With
`blockingRequirement` at 1 or 2, new (possibly desirable) pathfinding
behavior is possible, and it's about as fast as it was before the
optimizations in this latest round with IntDoubleOrderedMap and the
garbage reduction (but does work the old code didn't to do checks for
obstacles).