Some small improvements to Bresenham
This uses pooled 2D Coords when possible, allocates arrays of Coord that
are already the right size with line2D_ (instead of generating and then
converting from a Queue of Coord, which I have seen is a source of
slowdown in CoordPacker), and maintains the API smelC introduced in the
last commit. LOS now prefers the line2D_ method since it doesn't need to
generate a Queue twice, and can now generate an array once and feed
(part of) it into a Queue once.