Better support for arrays as OrderedMap/Set Keys
Before, the default `.hashCode()` method was called on arrays used as
keys in OrderedMap and OrderedSet (generally on primitive arrays), the
same as LinkedHashMap/Set. That unfortunately is an identity hash code,
and two different arrays with identical contents would hash differently
every time. Now, CrossHash can be used (and is in CoordPacker) to define
a hashing strategy that can be used in place of the default
`.hashCode()`. This is simply an extra arg to the constructor that is
usually a constant in CrossHash that defines a simple hashing strategy,
but on Java 8 and higher can even be a lambda, though I haven't tested
that yet. You can create a `DefaultHasher` that will just call
`.hashCode()` on its input if you want identity hash codes for arrays or
the normal `.hashCode()` behavior for non-arrays, and one of these is
used if no IHasher is specified in OrderedMap/Set.