Improve syllable counting in FakeLanguageGen
Previously, large closing sections of words (which are not necessarily
one syllable, such as the suffix "-iedly" in imitation English) threw
off the count for how many syllables a word actually was, and the whole
suffix counted as one syllable. This was especially problematic for
imitation Swahili, which used suffixes to handle redoubling, and could
produce 4-syllable redoubled words when a 2-syllable word was requested
(such as by LanguageCipher). This is handled better now, and some other
language stuff needed help too (Greek didn't have the possibility to use
the letter delta, for instance). This is all preparing for some
additions to FakeLanguageGen.