Major improvements to GreasedRegion
Now you can specify an amount of times to perform an expand, fringe,
retract, or surface operation, you can flood-fill a bounded area, you
can find the distinct (unconnected) regions of a GreasedRegion with
split(), and so on. There's also some more constructors to even out the
functionality relative to CoordPacker and its packSeveral() methods.
Width and height changed order in parameter lists on some of the methods
to match the rest of SquidLib. I don't think anyone is currently using
GreasedRegion, but that would be a breaking change if you used the
previous signature of a specific constructor, `GreasedRegion(boolean[]
bits, int height, int width)`.