New separatedPortion() method in GreasedRegion
This fills the role of `CoordPacker.fractionPacked()` but uses SobolQRNG
internally (the first time it's been used in SquidLib, to my knowledge)
to ensure the points are far enough spaced out most of the time. This
means it doesn't need to traverse the Hilbert Curve, like CoordPacker
does, to get its spacing-out done. There's also a new `fit()` method
that allows you to give two `double`s for x and y and find some kind of
fitting to an "on" cell in the GreasedRegion. Also, there's an addition
to the vulgarity filter in FakeLanguageGen.