Make a good subset of Sobol GWT-compatible
This is of course good for GreasedRegion, which is now GWT-compatible
again but not thread-safe (though it probably never was), and will allow
up to 16-dimensional points to be generated with SobolQRNG on GWT, or up
to 1000-dimensional points if GWT isn't needed. There's also a "be nice
to the GC" update that allows SobolQRNG's produced `double[]` values to
fill an existing `double[]` instead of allocating many new ones. The
SobolQRNG in GreasedRegion is now static, which shouldn't be a problem
since it doesn't deal with native resources (any more) and has the
mutable part of its state reset (but the object is not re-constructed)
before each group of calls.