Start using GreasedRegion. 106x speedup in SDG
The benchmark for SectionDungeonGenerator was done in two other
projects, one with the previous commit and one with this commit, and may
not be completely accurate. However, it's at least one order of
magnitude better, so I'm happy about it. GreasedRegion has a number of
new features here, OrderedMap has a new method, CoordPacker can produce
GreasedRegion values now, and on this laptop SectionDungeonGenerator
(with all the Placement info requested) generates 2348 dungeons in one
minute instead of before, which was 22 dungeons in one minute.